New working code

+++ src/main/java/com/boozeonwheel/product/domain/master/CategoryDTO.java
@@ -11,14 +11,14 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 })
 public class CategoryDTO {
 	@JsonProperty("categoryId")
-	private int categoryId;
+	private long categoryId;
 	
 	@JsonProperty("categoryName")
 	private String categoryName;
-	public int getCategoryId() {
+	public long getCategoryId() {
 		return categoryId;
 	}
-	public void setCategoryId(int categoryId) {
+	public void setCategoryId(long categoryId) {
 		this.categoryId = categoryId;
 	}
 	public String getCategoryName() {
